schemachanger: implement `OWNER TO` for domains
This change implements the `ALTER DOMAIN ... OWNER
TO` statement which was previously unsupported.
The legacy schemachanger did not support domain
types so logic tests are added to cover the
feature.

The `ALTER TYPE` implementation is refactored so
the `ALTER DOMAIN` implementation can use its
logic.

Closes: #166941
Epic: CRDB-62146

Release note (sql change): The `ALTER DOMAIN ...
OWNER TO` statement is supported.
